phase E: separate debug concerns from core compiler
Moved all DWARF/debug code out of module.rs into debug_info.rs:
- SrcLocEntry, FunctionDebugInfo, mir_type_to_var_info, compute_inline_ranges
- collect_function_debug() extracts srcloc collection from gen_function
- resolve_debug_info() extracts FuncId→SymbolId resolution from emit_object

module.rs now has a single call into debug_info after each compilation —
no DWARF types, no type mapping, no inline range computation.

Added PassContext to the MirPass trait so passes can write metadata
to a shared context instead of special-casing by name. The inlining
pass writes inline_regions directly to ctx, no string matching needed.
