+++ compiler/crates/rask-mir/src/transform/pass.rs
@@ -3,7 +3,7 @@
 //! Pass manager — runs MIR optimization passes in sequence.
 //!
 //! Each pass implements `MirPass`. The `PassManager` runs them in order,
-//! providing shared context for cross-function analysis.
+//! threading a `PassContext` through for side-channel metadata collection.
 
 use std::collections::HashMap;
 use crate::MirFunction;
@@ -14,26 +14,28 @@ pub trait MirPass {
     /// Short name for logging/debugging.
     fn name(&self) -> &str;
 
-    /// Run on the full set of functions. Default iterates per-function.
-    /// Override for cross-function passes (e.g., closure escape analysis).
-    fn run(&self, fns: &mut Vec<MirFunction>) {
+    /// Run on the full set of functions with shared context.
+    /// Default iterates per-function (ignoring ctx).
+    fn run(&self, fns: &mut Vec<MirFunction>, _ctx: &mut PassContext) {
         for func in fns.iter_mut() {
             self.run_function(func);
         }
     }
 
     /// Run on a single function. Default is no-op.
-    /// Most passes override this.
     fn run_function(&self, _func: &mut MirFunction) {}
 }
 
-/// Side-channel metadata collected during the pass pipeline.
-/// Keeps debug concerns out of MirFunction.
+/// Shared context threaded through the pass pipeline.
+/// Passes can write metadata here; downstream consumers read it.
 #[derive(Debug, Default)]
-pub struct PipelineResult {
+pub struct PassContext {
     /// DI5: inline region metadata per caller function name.
     pub inline_regions: HashMap<String, Vec<InlineRegion>>,
 }
+
+/// Convenience alias — the result of running the pipeline is the context.
+pub type PipelineResult = PassContext;
 
 /// Runs a sequence of MIR passes.
 pub struct PassManager {
@@ -50,18 +52,13 @@ impl PassManager {
         self.passes.push(Box::new(pass));
     }
 
-    /// Run all passes in order. Returns side-channel metadata from passes
-    /// that produce debug info (e.g., inlining).
+    /// Run all passes in order. Returns the accumulated context.
     pub fn run(&self, fns: &mut Vec<MirFunction>) -> PipelineResult {
-        let mut result = PipelineResult::default();
+        let mut ctx = PassContext::default();
         for pass in &self.passes {
-            if pass.name() == "inlining" {
-                result.inline_regions = crate::transform::inline::inline_functions(fns);
-            } else {
-                pass.run(fns);
-            }
+            pass.run(fns, &mut ctx);
         }
-        result
+        ctx
     }
 
     /// Build the default optimization pipeline.
@@ -88,7 +85,7 @@ pub struct ClosureOptimizationPass;
 
 impl MirPass for ClosureOptimizationPass {
     fn name(&self) -> &str { "closure_optimization" }
-    fn run(&self, fns: &mut Vec<MirFunction>) {
+    fn run(&self, fns: &mut Vec<MirFunction>, _ctx: &mut PassContext) {
         crate::optimize_all_closures(fns);
     }
 }
@@ -98,11 +95,8 @@ pub struct InliningPass;
 
 impl MirPass for InliningPass {
     fn name(&self) -> &str { "inlining" }
-    // Inlining is dispatched directly by PassManager::run to capture
-    // the returned InlineRegion metadata. This run() is unused but
-    // kept for the trait contract.
-    fn run(&self, fns: &mut Vec<MirFunction>) {
-        let _ = crate::transform::inline::inline_functions(fns);
+    fn run(&self, fns: &mut Vec<MirFunction>, ctx: &mut PassContext) {
+        ctx.inline_regions = crate::transform::inline::inline_functions(fns);
     }
 }
 
@@ -111,7 +105,7 @@ pub struct StringConcatPass;
 
 impl MirPass for StringConcatPass {
     fn name(&self) -> &str { "string_concat" }
-    fn run(&self, fns: &mut Vec<MirFunction>) {
+    fn run(&self, fns: &mut Vec<MirFunction>, _ctx: &mut PassContext) {
         crate::optimize_string_concat(fns);
     }
 }
@@ -121,7 +115,7 @@ pub struct CloneElisionPass;
 
 impl MirPass for CloneElisionPass {
     fn name(&self) -> &str { "clone_elision" }
-    fn run(&self, fns: &mut Vec<MirFunction>) {
+    fn run(&self, fns: &mut Vec<MirFunction>, _ctx: &mut PassContext) {
         crate::elide_clones(fns);
     }
 }
@@ -161,7 +155,7 @@ pub struct GenerationCoalescingPass;
 
 impl MirPass for GenerationCoalescingPass {
     fn name(&self) -> &str { "generation_coalescing" }
-    fn run(&self, fns: &mut Vec<MirFunction>) {
+    fn run(&self, fns: &mut Vec<MirFunction>, _ctx: &mut PassContext) {
         crate::coalesce_generation_checks(fns);
     }
 }
